refactor ARDetectView
