+++ PlantApp/Views/AR Detect View/ARDetectViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  ARDetectViewModel.swift
+//  PlantApp
+//
+//  Created by Aeliana Shen on 3/22/24.
+//
+
+import Foundation
+import SwiftUI
+
+class ARDetectViewModel: ObservableObject {
+    let guidelineString = "Open your camera and decide where to place your plant. We are going to find the best plant choice for you!"
+    let buttonString = "Find a place for your plant"
+}
